将foundation目录下的接口放到a2a并适配主流程
