+++ applications/a2a_service/.env.example
@@ -51,17 +51,6 @@ FASTAPI_WORKERS=1
 LOG_LEVEL=INFO
 LOG_DIR=logs
 
-# ── Runtime DB（DB 权威 + Redis 缓存）──────────────────────────────────────────
-# 开关：true=启用DB持久化（先写DB再写Redis，Redis miss时回源DB）；false=纯Redis模式（不连DB）
-# 禁用DB时以下字段允许留空
-RUNTIME_DB_ENABLED=false
-RUNTIME_DB_TYPE=gaussdb
-RUNTIME_DB_HOST=127.0.0.1
-RUNTIME_DB_PORT=8000
-RUNTIME_DB_NAME=postgres
-RUNTIME_DB_USER=root
-RUNTIME_DB_PASSWORD=
-
 # ── LLM（上游保留字段，供其他 Agent 参考；EDPAgent 不读此组）──────────────
 LLM_PROVIDER=llm_OpenAI
 LLM_API_BASE=https://api.deepseek.com
@@ -213,3 +202,13 @@ ACTIVE_SCENARIO=AgentRule_Wealth_purchase
 #   - filtered：按场景配置只注册相关 Skill（默认，推荐）
 #   - all：注册所有 Skill（向后兼容模式，开发调试时使用）
 DPA_SKILL_REGISTRATION_MODE=filtered
+
+# ── Runtime DB（DB 持久化开关）────────────────────────────────────────────
+# true=启用DB持久化（先写DB再写Redis，Redis miss时回源DB）；false=纯Redis模式（不连DB，和勇哥代码一致）
+RUNTIME_DB_ENABLED=false
+RUNTIME_DB_TYPE=gaussdb
+RUNTIME_DB_HOST=127.0.0.1
+RUNTIME_DB_PORT=8000
+RUNTIME_DB_NAME=postgres
+RUNTIME_DB_USER=root
+RUNTIME_DB_PASSWORD=
