+++ applications/a2a_service/common/schema_manager.py
